Database Integrity Exception Implemented.

+++ src/main/java/com/devsuperior/_dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior._dscommerce.services;
 import com.devsuperior._dscommerce.dto.ProductDTO;
 import com.devsuperior._dscommerce.entities.Product;
 import com.devsuperior._dscommerce.repositories.ProductRepository;
+import com.devsuperior._dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior._dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,8 +25,10 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Optional<Product> result = repository.findById(id);
-        return new ProductDTO(result.get());
+
+        Product result = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado."));
+        return new ProductDTO(result);
 
     }
 
@@ -49,21 +56,36 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
 
-        // Essa operação não vai ao BD, ela acessa os valores de um objeto pelo mapeamento JPA feito
-        Product entity = repository.getReferenceById(id);
+        try{
+            // getReferenceById(): Essa operação não vai ao BD, ela acessa os valores de um
+            // objeto pelo mapeamento JPA feito
+            Product entity = repository.getReferenceById(id);
 
-        copyDtoToEntity(dto, entity);
+            copyDtoToEntity(dto, entity);
 
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
 
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
-
+    // SUPPORTS: só executa a transação se o método estiver no contexto de outra transação
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
+
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
